studentCard replaced with the new one
