--- src/styles/new_card.scss
@@ -1,114 +1,0 @@
-.new_card {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: center;
-  margin-top: 6rem;
-  @include tablet-only {
-    margin-top: 4rem;
-  }
-  @include mobile {
-    margin-top: 2rem;
-  }
-
-  &-details {
-    flex: 2 1 240px;
-    @include tablet {
-      text-align: right;
-    }
-  }
-  &-quote {
-    flex: 3 2 240px;
-    text-align: center;
-  }
-  &-image {
-    width: 5rem;
-    height: 5rem;
-    border-radius: 100%;
-  }
-  &-profile {
-    display: inline-flex;
-    align-items: center;
-    margin: 0;
-    & > img {
-      margin-right: 1rem;
-    }
-    @include tablet {
-      flex-direction: row-reverse;
-      & > img{
-        margin-right: 0;
-        margin-left: 1rem;
-      }
-    }
-  }
-  &-division {
-    background-color: $grey-lighter;
-    &.d-vertical {
-      display: flex;
-      align-self: stretch;
-      margin: 0 1rem;
-      width: 1px;
-      height: auto;
-    }
-    &.d-horizontal {
-      margin-top: 2rem;
-      width: 100%;
-      height: 1px;
-    }
-  }
-  &
-  &-name {
-    font-size: 1.5rem;
-    font-weight: 600;
-    line-height: 1;
-  }
-  &-quote {
-    position: relative;
-    padding: 2rem 0;
-    .quote {
-      position: absolute;
-      line-height: 1rem;
-      color: $gray-light;
-      font-family: 'Playfair Display', serif;
-      font-size: 6rem;
-      font-weight: 200;
-      &.first {
-        top: 2rem;
-        left: 0px;
-      }
-      &.last {
-        bottom: 0px;
-        right: 0px;
-      }
-    }
-    &-content {
-      font-size: 1.25rem;
-      line-height: 1.5;
-      font-style: italic;
-      color: $gray-dark;
-      margin: 1.5rem;
-      @include mobile {
-        margin: 1rem;
-      }
-    }
-  }
-  &-link,
-  &-link:hover {
-    color: $gray-dark;
-    display: inline-flex;
-    align-items: center;
-  }
-  &-field {
-    margin-bottom: .75rem;
-    & > a > h5 {
-      margin-bottom: 0;
-      margin-right: .5rem;
-    }
-  }
-  &-label {
-    font-weight: 500;
-    color: $gray-dark;
-    margin-bottom: .25rem;
-  }
-  
-}
